Added "oldway" to MyBatis101. Start creating blog example.
The MyBatis101 was changed because it did not properly close the session
and it only showed the new mapper way, not the older mapper way to query.
It also was wrong in that it didn't actually accept a param in the mapping
xml.  Fixed that.

Blog is currently failing due to mapping errors -> not sure why.
Need to compare to the MyBatis101 code and get it working before proceeding.